feat(3.3): change color of light over time
Lighting -> Materials -> Setting materials -> Different light colors

Think of lights in disco club. The object can only reflect or absorb the colors the light shining at it contains. If light contains no red, green and/or blue value, object will not reflect that component of RGB model, as it's simply not available. As evidenced in this iteration of "Materials" lecture, when light source shines an exclusively green light at object, perceived color of object becomes green. Same is true for exclusively red and blue light.