feat(3.5): model light source as a spotlight one
Lighting -> Light casters -> Spotlight -> Flashlight

Reference: https://wiki.ogre3d.org/tiki-index.php?page=-Point+Light+Attenuation

Spotlight source is a light source with a given position in world space that illuminates light rays ONLY IN A SPECIFIC DIRECTION, with its light rays fading out over distance. That means that only the objects within a certain radius of the spotlight's direction are lit and everything else stays dark. A spotlight in OpenGL is represented by: a world-space position, a direction and a cutoff angle that specifies the radius of the spotlight. I set the light to cover the distance of 50.0f units, so parameters of attenuation are set to ones specified in a corresponding row of table at Ogre3D's wiki.